Fixing format errors

+++ ivy/functional/backends/jax/creation.py
@@ -124,9 +124,9 @@ def asarray(
         and len(object_in) != 0
         and dtype is None
     ):
-        dtype = default_dtype(item=object_in,as_native_dtype=True)
+        dtype = default_dtype(item=object_in, as_native_dtype=True)
         if copy is True:
-            return to_dev(jnp.array(object_in, dtype = dtype, copy=True), device=device)
+            return to_dev(jnp.array(object_in, dtype=dtype, copy=True), device=device)
         else:
             return to_dev(jnp.asarray(object_in, dtype=dtype), device=device)
     else:
