+++ ivy/functional/backends/tensorflow/creation.py
@@ -27,13 +27,13 @@ def asarray(object_in, dtype=None, device=None, copy=None):
             if dtype is None and not isinstance(object_in, tf.Tensor):
                 try:
                     dtype = default_dtype(item=object_in, as_native=True)
-                    tensor = tf.convert_to_tensor(object_in,dtype= dtype)
+                    tensor = tf.convert_to_tensor(object_in, dtype=dtype)
                 except (TypeError, ValueError):
-                    dtype = default_dtype(dtype, object_in,True)
+                    dtype = default_dtype(dtype, object_in, True)
                     tensor = tf.convert_to_tensor(
-                            ivy.nested_map(object_in, lambda x: tf.cast(x, dtype)),
-                            dtype=dtype,
-                        )
+                        ivy.nested_map(object_in, lambda x: tf.cast(x, dtype)),
+                        dtype=dtype,
+                    )
                 return tf.identity(tf.cast(tensor, dtype))
             else:
                 dtype = as_ivy_dtype(default_dtype(dtype, object_in))
